Reorganize tests and get all passing
