+++ tests/test_datasets.py
@@ -1,3 +1,5 @@
+"""Tests for dataset loading."""
+
 import pickle
 import sys
 from dataclasses import dataclass
